+++ Tests/HoneybadgerTests/main.m
@@ -3,6 +3,7 @@
 // One declaration + one call per suite. Later tasks append theirs here.
 void run_smoke_tests(void);
 void run_environment_tests(void);
+void run_endpoint_tests(void);
 void run_payload_tests(void);
 void run_pending_report_tests(void);
 void run_binary_image_tests(void);
@@ -17,6 +18,7 @@ int main(int argc, char** argv)
     @autoreleasepool {
         run_smoke_tests();
         run_environment_tests();
+        run_endpoint_tests();
         run_payload_tests();
         run_pending_report_tests();
         run_binary_image_tests();